+++ src/components/Span/index.ts
@@ -1,0 +1,3 @@
+import RvSpan from "./RvSpan";
+
+export { RvSpan };