+++ src/components/Span/style/index.ts
@@ -1,0 +1,3 @@
+import StyledSpan from "./base";
+
+export { StyledSpan };