+++ src/types/components/index.ts
@@ -1,3 +1,4 @@
 export * from "./Div";
 export * from "./Button";
+export * from "./Span";
 export * from "../CSSProperties";