+++ CHANGELOG.md
@@ -1,3 +1,12 @@
+## 0.3.17 (2023/6/1)
+
+-   ### Fix Add `;` at the end of the property (style/\*.ts)
+-   ### Add Modified RvButton to be able to set icon.
+-   ### Fix Fixed `StyledIconButton` to accept each property as an argument.
+-   ### Add Add Type
+-   ### Fix Change variable names for each Type
+-   ### Create Create `RvSpan` Components
+
 ## 0.2.17 (2023/05/31)
 
 -   ### Create RvButton and RvIconButton(Bata) Components